+++ Sources/KeyboardKit/Callouts/Providers/BaseCalloutActionProvider.swift
@@ -1,5 +1,5 @@
 //
-//  BaseSecondaryCalloutActionProvider.swift
+//  BaseCalloutActionProvider.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2021-02-01.