+++ Sources/KeyboardKit/Callouts/Providers/CalloutActionProvider.swift
@@ -10,7 +10,7 @@ import Foundation
 
 /**
  This protocol can be implemented by any classes that can be
- used to get secondary callout actions for a keyboard action.
+ used to get callout actions for a keyboard action.
  
  KeyboardKit automatically creates an implementation of this
  protocol and binds it to ``KeyboardInputViewController``.