+++ Sources/KeyboardKit/Views/System/SystemKeyboard.swift
@@ -78,7 +78,7 @@ public struct SystemKeyboard<ButtonView: View>: View {
     public typealias KeyboardItemWidth = CGFloat
     
     private var actionCalloutStyle: ActionCalloutStyle {
-        var style = appearance.secondaryInputCalloutStyle()
+        var style = appearance.actionCalloutStyle()
         let insets = layoutConfig.buttonInsets
         style.callout.buttonInset = CGSize(width: insets.leading, height: insets.top)
         return style