+++ Sources/KeyboardKit/_Deprecated/KeyboardAppearance+Deprecated.swift
@@ -2,6 +2,7 @@ import Foundation
 
 public extension KeyboardAppearance {
     
+    @available(*, deprecated, renamed: "actionCalloutStyle")
     func secondaryInputCalloutStyle() -> ActionCalloutStyle {
         actionCalloutStyle()
     }