+++ Sources/KeyboardKit/_Deprecated/View+SecondaryInputCallout.swift
@@ -2,14 +2,6 @@ import SwiftUI
 
 public extension View {
     
-    /**
-     This modifier can be applied to any view that should be
-     able to present a secondary input callout.
-     
-     - Parameters:
-       - context: The context to bind against.
-       - style: The style to apply to the view, by default `.standard`.
-     */
     @available(*, deprecated, renamed: "actionCallout")
     func secondaryInputCallout(
         context: ActionCalloutContext,