+++ RELEASE_NOTES.md
@@ -26,6 +26,7 @@ This version renames some types and init parameters to make things tighter in pr
 * `CalloutActionProvider` `secondaryCalloutActions` has been renamed to `calloutActions`.
 * `DisabledSecondaryCalloutActionProvider` has been renamed to `DisabledCalloutActionProvider`.
 * `EnglishSecondaryCalloutActionProvider` has been renamed to `EnglishCalloutActionProvider`.
+* `KeyboardInputViewController` `keyboardInputCalloutContext` has been renamed to `inputCalloutContext`.
 * `KeyboardInputViewController` `keyboardSecondaryInputCalloutContext` has been renamed to `actionCalloutContext`.
 * `KeyboardInputViewController` `keyboardSecondaryCalloutActionProvider` has been renamed to `calloutActionContext`.
 * `LocalizedSecondaryCalloutActionProvider` has been renamed to `LocalizedCalloutActionProvider`.